refactor, bring code into 'changeset' directory
