+++ integration-tests/deployment/ccip/changeset/add_chain_test.go
@@ -1,6 +1,7 @@
-package ccipdeployment
+package changeset
 
 import (
+	"github.com/smartcontractkit/chainlink/integration-tests/deployment/ccip"
 	"testing"
 	"time"
 
@@ -27,8 +28,8 @@ import (
 
 func TestAddChainInbound(t *testing.T) {
 	// 4 chains where the 4th is added after initial deployment.
-	e := NewMemoryEnvironmentWithJobs(t, logger.TestLogger(t), 4, 4)
-	state, err := LoadOnchainState(e.Env, e.Ab)
+	e := ccipdeployment.NewMemoryEnvironmentWithJobs(t, logger.TestLogger(t), 4, 4)
+	state, err := ccipdeployment.LoadOnchainState(e.Env, e.Ab)
 	require.NoError(t, err)
 	// Take first non-home chain as the new chain.
 	newChain := e.Env.AllChainSelectorsExcluding([]uint64{e.HomeChainSel})[0]
@@ -36,54 +37,54 @@ func TestAddChainInbound(t *testing.T) {
 	initialDeploy := e.Env.AllChainSelectorsExcluding([]uint64{newChain})
 
 	feeds := state.Chains[e.FeedChainSel].USDFeeds
-	tokenConfig := NewTokenConfig()
-	tokenConfig.UpsertTokenInfo(LinkSymbol,
+	tokenConfig := ccipdeployment.NewTokenConfig()
+	tokenConfig.UpsertTokenInfo(ccipdeployment.LinkSymbol,
 		pluginconfig.TokenInfo{
-			AggregatorAddress: cciptypes.UnknownEncodedAddress(feeds[LinkSymbol].Address().String()),
-			Decimals:          LinkDecimals,
+			AggregatorAddress: cciptypes.UnknownEncodedAddress(feeds[ccipdeployment.LinkSymbol].Address().String()),
+			Decimals:          ccipdeployment.LinkDecimals,
 			DeviationPPB:      cciptypes.NewBigIntFromInt64(1e9),
 		},
 	)
-	tokenConfig.UpsertTokenInfo(WethSymbol,
+	tokenConfig.UpsertTokenInfo(ccipdeployment.WethSymbol,
 		pluginconfig.TokenInfo{
-			AggregatorAddress: cciptypes.UnknownEncodedAddress(feeds[WethSymbol].Address().String()),
-			Decimals:          WethDecimals,
+			AggregatorAddress: cciptypes.UnknownEncodedAddress(feeds[ccipdeployment.WethSymbol].Address().String()),
+			Decimals:          ccipdeployment.WethDecimals,
 			DeviationPPB:      cciptypes.NewBigIntFromInt64(1e9),
 		},
 	)
-	err = DeployCCIPContracts(e.Env, e.Ab, DeployCCIPContractConfig{
+	err = ccipdeployment.DeployCCIPContracts(e.Env, e.Ab, ccipdeployment.DeployCCIPContractConfig{
 		HomeChainSel:       e.HomeChainSel,
 		FeedChainSel:       e.FeedChainSel,
 		ChainsToDeploy:     initialDeploy,
 		TokenConfig:        tokenConfig,
-		MCMSConfig:         NewTestMCMSConfig(t, e.Env),
+		MCMSConfig:         ccipdeployment.NewTestMCMSConfig(t, e.Env),
 		FeeTokenContracts:  e.FeeTokenContracts,
 		CapabilityRegistry: state.Chains[e.HomeChainSel].CapabilityRegistry.Address(),
 		OCRSecrets:         deployment.XXXGenerateTestOCRSecrets(),
 	})
 	require.NoError(t, err)
-	state, err = LoadOnchainState(e.Env, e.Ab)
+	state, err = ccipdeployment.LoadOnchainState(e.Env, e.Ab)
 	require.NoError(t, err)
 
 	// Connect all the existing lanes.
 	for _, source := range initialDeploy {
 		for _, dest := range initialDeploy {
 			if source != dest {
-				require.NoError(t, AddLane(e.Env, state, source, dest))
+				require.NoError(t, ccipdeployment.AddLane(e.Env, state, source, dest))
 			}
 		}
 	}
 
-	rmnHomeAddress, err := deployment.SearchAddressBook(e.Ab, e.HomeChainSel, RMNHome)
+	rmnHomeAddress, err := deployment.SearchAddressBook(e.Ab, e.HomeChainSel, ccipdeployment.RMNHome)
 	require.NoError(t, err)
 	require.True(t, common.IsHexAddress(rmnHomeAddress))
 	rmnHome, err := rmn_home.NewRMNHome(common.HexToAddress(rmnHomeAddress), e.Env.Chains[e.HomeChainSel].Client)
 	require.NoError(t, err)
 
 	//  Deploy contracts to new chain
-	err = DeployChainContracts(e.Env, e.Env.Chains[newChain], e.Ab, e.FeeTokenContracts[newChain], NewTestMCMSConfig(t, e.Env), rmnHome)
+	err = ccipdeployment.DeployChainContracts(e.Env, e.Env.Chains[newChain], e.Ab, e.FeeTokenContracts[newChain], ccipdeployment.NewTestMCMSConfig(t, e.Env), rmnHome)
 	require.NoError(t, err)
-	state, err = LoadOnchainState(e.Env, e.Ab)
+	state, err = ccipdeployment.LoadOnchainState(e.Env, e.Ab)
 	require.NoError(t, err)
 
 	// Transfer onramp/fq ownership to timelock.
@@ -116,12 +117,12 @@ func TestAddChainInbound(t *testing.T) {
 	_, err = deployment.ConfirmIfNoError(e.Env.Chains[e.HomeChainSel], tx, err)
 	require.NoError(t, err)
 
-	acceptOwnershipProposal, err := GenerateAcceptOwnershipProposal(state, e.HomeChainSel, initialDeploy)
+	acceptOwnershipProposal, err := ccipdeployment.GenerateAcceptOwnershipProposal(state, e.HomeChainSel, initialDeploy)
 	require.NoError(t, err)
-	acceptOwnershipExec := SignProposal(t, e.Env, acceptOwnershipProposal)
+	acceptOwnershipExec := ccipdeployment.SignProposal(t, e.Env, acceptOwnershipProposal)
 	// Apply the accept ownership proposal to all the chains.
 	for _, sel := range initialDeploy {
-		ExecuteProposal(t, e.Env, acceptOwnershipExec, state, sel)
+		ccipdeployment.ExecuteProposal(t, e.Env, acceptOwnershipExec, state, sel)
 	}
 	for _, chain := range initialDeploy {
 		owner, err2 := state.Chains[chain].OnRamp.Owner(nil)
@@ -141,9 +142,9 @@ func TestAddChainInbound(t *testing.T) {
 	// Generate and sign inbound proposal to new 4th chain.
 	chainInboundProposal, err := NewChainInboundProposal(e.Env, state, e.HomeChainSel, newChain, initialDeploy)
 	require.NoError(t, err)
-	chainInboundExec := SignProposal(t, e.Env, chainInboundProposal)
+	chainInboundExec := ccipdeployment.SignProposal(t, e.Env, chainInboundProposal)
 	for _, sel := range initialDeploy {
-		ExecuteProposal(t, e.Env, chainInboundExec, state, sel)
+		ccipdeployment.ExecuteProposal(t, e.Env, chainInboundExec, state, sel)
 	}
 	// TODO This currently is not working - Able to send the request here but request gets stuck in execution
 	// Send a new message and expect that this is delivered once the chain is completely set up as inbound
@@ -153,28 +154,28 @@ func TestAddChainInbound(t *testing.T) {
 	addDonProp, err := AddDonAndSetCandidateProposal(state, e.Env, nodes, deployment.XXXGenerateTestOCRSecrets(), e.HomeChainSel, e.FeedChainSel, newChain, tokenConfig, types.PluginTypeCCIPCommit)
 	require.NoError(t, err)
 
-	addDonExec := SignProposal(t, e.Env, addDonProp)
-	ExecuteProposal(t, e.Env, addDonExec, state, e.HomeChainSel)
+	addDonExec := ccipdeployment.SignProposal(t, e.Env, addDonProp)
+	ccipdeployment.ExecuteProposal(t, e.Env, addDonExec, state, e.HomeChainSel)
 
 	t.Logf("Executing promote candidate proposal for exec plugin on chain %d", newChain)
 	setCandidateForExecProposal, err := SetCandidatePluginProposal(state, e.Env, nodes, deployment.XXXGenerateTestOCRSecrets(), e.HomeChainSel, e.FeedChainSel, newChain, tokenConfig, types.PluginTypeCCIPExec)
 	require.NoError(t, err)
-	setCandidateForExecExec := SignProposal(t, e.Env, setCandidateForExecProposal)
-	ExecuteProposal(t, e.Env, setCandidateForExecExec, state, e.HomeChainSel)
+	setCandidateForExecExec := ccipdeployment.SignProposal(t, e.Env, setCandidateForExecProposal)
+	ccipdeployment.ExecuteProposal(t, e.Env, setCandidateForExecExec, state, e.HomeChainSel)
 
 	t.Logf("Executing promote candidate proposal for both commit and exec plugins on chain %d", newChain)
-	donPromoteProposal, err := PromoteCandidateProposal(state, e.HomeChainSel, newChain, nodes)
+	donPromoteProposal, err := PromoteAllCandidatesProposal(state, e.HomeChainSel, newChain, nodes)
 	require.NoError(t, err)
-	donPromoteExec := SignProposal(t, e.Env, donPromoteProposal)
-	ExecuteProposal(t, e.Env, donPromoteExec, state, e.HomeChainSel)
+	donPromoteExec := ccipdeployment.SignProposal(t, e.Env, donPromoteProposal)
+	ccipdeployment.ExecuteProposal(t, e.Env, donPromoteExec, state, e.HomeChainSel)
 
 	// verify if the configs are updated
-	require.NoError(t, ValidateCCIPHomeConfigSetUp(
+	require.NoError(t, ccipdeployment.ValidateCCIPHomeConfigSetUp(
 		state.Chains[e.HomeChainSel].CapabilityRegistry,
 		state.Chains[e.HomeChainSel].CCIPHome,
 		newChain,
 	))
-	replayBlocks, err := LatestBlocksByChain(testcontext.Get(t), e.Env.Chains)
+	replayBlocks, err := ccipdeployment.LatestBlocksByChain(testcontext.Get(t), e.Env.Chains)
 	require.NoError(t, err)
 
 	// Now configure the new chain using deployer key (not transferred to timelock yet).
@@ -192,9 +193,9 @@ func TestAddChainInbound(t *testing.T) {
 	_, err = deployment.ConfirmIfNoError(e.Env.Chains[newChain], tx, err)
 	require.NoError(t, err)
 	// Set the OCR3 config on new 4th chain to enable the plugin.
-	latestDON, err := LatestCCIPDON(state.Chains[e.HomeChainSel].CapabilityRegistry)
+	latestDON, err := ccipdeployment.LatestCCIPDON(state.Chains[e.HomeChainSel].CapabilityRegistry)
 	require.NoError(t, err)
-	ocrConfigs, err := BuildSetOCR3ConfigArgs(latestDON.Id, state.Chains[e.HomeChainSel].CCIPHome, newChain)
+	ocrConfigs, err := ccipdeployment.BuildSetOCR3ConfigArgs(latestDON.Id, state.Chains[e.HomeChainSel].CCIPHome, newChain)
 	require.NoError(t, err)
 	tx, err = state.Chains[newChain].OffRamp.SetOCR3Configs(e.Env.Chains[newChain].DeployerKey, ocrConfigs)
 	require.NoError(t, err)
@@ -202,7 +203,7 @@ func TestAddChainInbound(t *testing.T) {
 	require.NoError(t, err)
 
 	// Assert the inbound lanes to the new chain are wired correctly.
-	state, err = LoadOnchainState(e.Env, e.Ab)
+	state, err = ccipdeployment.LoadOnchainState(e.Env, e.Ab)
 	require.NoError(t, err)
 	for _, chain := range initialDeploy {
 		cfg, err2 := state.Chains[chain].OnRamp.GetDestChainConfig(nil, newChain)
@@ -217,25 +218,25 @@ func TestAddChainInbound(t *testing.T) {
 	}
 	// Ensure job related logs are up to date.
 	time.Sleep(30 * time.Second)
-	ReplayLogs(t, e.Env.Offchain, replayBlocks)
+	ccipdeployment.ReplayLogs(t, e.Env.Offchain, replayBlocks)
 
 	// TODO: Send via all inbound lanes and use parallel helper
 	// Now that the proposal has been executed we expect to be able to send traffic to this new 4th chain.
 	latesthdr, err := e.Env.Chains[newChain].Client.HeaderByNumber(testcontext.Get(t), nil)
 	require.NoError(t, err)
 	startBlock := latesthdr.Number.Uint64()
-	seqNr := SendRequest(t, e.Env, state, initialDeploy[0], newChain, true)
+	seqNr := ccipdeployment.SendRequest(t, e.Env, state, initialDeploy[0], newChain, true)
 	require.NoError(t,
-		ConfirmCommitWithExpectedSeqNumRange(t, e.Env.Chains[initialDeploy[0]], e.Env.Chains[newChain], state.Chains[newChain].OffRamp, &startBlock, cciptypes.SeqNumRange{
+		ccipdeployment.ConfirmCommitWithExpectedSeqNumRange(t, e.Env.Chains[initialDeploy[0]], e.Env.Chains[newChain], state.Chains[newChain].OffRamp, &startBlock, cciptypes.SeqNumRange{
 			cciptypes.SeqNum(1),
 			cciptypes.SeqNum(seqNr),
 		}))
 	require.NoError(t,
-		ConfirmExecWithSeqNr(t, e.Env.Chains[initialDeploy[0]], e.Env.Chains[newChain], state.Chains[newChain].OffRamp, &startBlock, seqNr))
+		ccipdeployment.ConfirmExecWithSeqNr(t, e.Env.Chains[initialDeploy[0]], e.Env.Chains[newChain], state.Chains[newChain].OffRamp, &startBlock, seqNr))
 
 	linkAddress := state.Chains[newChain].LinkToken.Address()
 	feeQuoter := state.Chains[newChain].FeeQuoter
 	timestampedPrice, err := feeQuoter.GetTokenPrice(nil, linkAddress)
 	require.NoError(t, err)
-	require.Equal(t, MockLinkPrice, timestampedPrice.Value)
+	require.Equal(t, ccipdeployment.MockLinkPrice, timestampedPrice.Value)
 }
